requirements.txt
flask==2.3.3
authlib==1.2.1
python-dotenv==1.0.0
mysql-connector-python==8.0.32
requests==2.31.0
google-adk==0.1.0
google-generativeai==0.3.1 